Home 🏠

Functions wb_countries wb_topics wb_sources wb_regions wb_income_levels wb_lending_types wb_languages wb_indicators wb_cache wb_search wb_country_coverage wb_data

Vignettes 02 - Setup

News Changelog

wb_cache {wbstats}	R Documentation

Download an updated list of country, indicator, and source information

Description

Download an updated list of information regarding countries, indicators, sources, regions, indicator topics, lending types, income levels, and supported languages from the World Bank API

Usage

wb_cache(lang)

Arguments

lang 	 Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Value

A list containing the following items:

countries: The result of calling wb_countries()

indicators: The result of calling wb_indicators()

sources: The result of calling wb_sources()

topics: The result of calling wb_topics()

regions: The result of calling wb_regions()

income_levels: The result of calling wb_income_levels()

lending_types: The result of calling wb_lending_types()

languages: The result of calling wb_languages()