World Bank Information End Points

These functions are simple wrappers around the various useful API end points that are helpful for finding avaiable data and filtering the data you are interested in when using wb_data()

wb_countries(lang)

wb_topics(lang)

wb_sources(lang)

wb_regions(lang)

wb_income_levels(lang)

wb_lending_types(lang)

wb_languages()

Arguments

lang
Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Value

A tibble of information about the end point

See also