Home 🏠

Functions wb_countries wb_topics wb_sources wb_regions wb_income_levels wb_lending_types wb_languages wb_indicators wb_cache wb_search wb_country_coverage wb_data

Vignettes 02 - Setup

News Changelog

wb_indicators {wbstats}	R Documentation

Download Avialable Indicators from the World Bank

Description

This function returns a data.frame of indicator IDs and related information that are available for download from the World Bank API

Usage

wb_indicators(lang, include_archive = FALSE)

Arguments

include_archive 	 logical. If TRUE indicators that have been archived by the World Bank will be included in the return. Data for these additional indicators are not available through the standard API and querying them using wb_data() will not return data. Default is FALSE.

lang 	 Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Examples

# can get a new list of available indicators by downloading new cache

fresh_cache <- wb_cache()

fresh_indicators <- fresh_cache$indicators

# or by running the wb_indicators() function directly

fresh_indicators <- wb_indicators()

# include archived indicators
# see include_archive parameter description

indicators_with_archive <- wb_indicators(include_archive = TRUE)