Home 🏠

Functions wb_countries wb_topics wb_sources wb_regions wb_income_levels wb_lending_types wb_languages wb_indicators wb_cache wb_search wb_data

Vignettes 02 - Setup

News Changelog

wb_search {wbstats}	R Documentation

Search indicator information available through the World Bank API

Description

This function allows finds indicators that match a search term and returns a data frame of matching results

Usage

wb_search(
  pattern,
  fields = c("indicator_id", "indicator", "indicator_desc"),
  extra = FALSE,
  cache,
  ignore.case = TRUE,
  ...
)

Arguments

pattern 	 Character string or regular expression to be matched

fields 	 Character vector of column names through which to search

extra 	 if FALSE, only the indicator ID and short name are returned, if TRUE, all columns of the cache parameter's indicators data frame are returned. Default is FALSE

ignore.case 	 if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching. Default is TRUE

cache 	 List of data frames returned from wb_cache(). If omitted, wb_cachelist is used

... 	 Any additional grep() agruments you which to pass

Value

a data.table/data.frame with indicators that match the search pattern.

Examples

d <- wb_search(pattern = "education")

d <- wb_search(pattern = "Food and Agriculture Organization", fields = "source_org")

# with regular expression operators
# 'poverty' OR 'unemployment' OR 'employment'
d <- wb_search(pattern = "poverty|unemployment|employment")

# pass any other grep argument along as well
# everything without 'education'
d <- wb_search(pattern = "education", invert = TRUE)

# contains "gdp" AND "trade"
d <- wb_search("^(?=.*gdp)(?=.*trade).*", perl = TRUE)

# contains "gdp" and NOT "trade"
d <- wb_search("^(?=.*gdp)(?!.*trade).*", perl = TRUE)