Home 🏠

Functions wb_countries wb_topics wb_sources wb_regions wb_income_levels wb_lending_types wb_languages wb_indicators wb_cache wb_search wb_data

Vignettes 02 - Setup

News Changelog

wb_end_point_info {wbstats}	R Documentation

World Bank Information End Points

Description

These functions are simple wrappers around the various useful API end points that are helpful for finding avaiable data and filtering the data you are interested in when using wb_data()

Usage

wb_countries(lang)
wb_topics(lang)
wb_sources(lang)
wb_regions(lang)
wb_income_levels(lang)
wb_lending_types(lang)
wb_languages()

Arguments

lang 	 Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Value

A data.table/data.frame of information about the end point. A data.table is used instead of a plain data.frame so that large tables (e.g. the ~28k rows returned by wb_indicators()) print truncated to a handful of rows instead of flooding the console.

See Also

wb_cache()