Download updated data source information from World Bank API

Download updated information on available data sources from the World Bank API

wbsources(lang = c("en", "es", "fr", "ar", "zh"))

Arguments

lang
Language in which to return the results. If lang is unspecified, english is the default.

Value

A data frame of available data scources with related information

Note

Not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA. For an enumeration of supported languages by data source please see wbdatacatalog. The options for lang are:

en: English

es: Spanish

fr: French

ar: Arabic

zh: Mandarin

Examples

wbsources()

#>    sourceID lastUpdated
#> 1         1  2019-10-23
#> 2         2  2019-12-20
#> 3         3  2019-11-07
#> 4         5  2016-03-21
#> 5         6  2019-12-02
#> 6        11  2013-02-22
#> 7        12  2019-12-10
#> 8        13  2017-01-03
#> 9        14  2019-12-20
#> 10       15  2019-09-13
#> 11       16  2019-12-20
#> 12       18  2019-12-23
#> 13       19  2018-09-19
#> 14       20  2020-01-27
#> 15       22  2020-01-31
#> 16       23  2020-01-31
#> 17       24  2019-09-24
#> 18       25  2019-12-20
#> 19       27  2020-01-08
#> 20       28  2018-10-15
#> 21       29  2019-05-02
#> 22       30  2016-03-31
#> 23       31  2019-08-22
#> 24       32  2019-10-30
#> 25       33  2019-02-27
#> 26       34  2013-04-12
#> 27       35  2018-06-30
#> 28       36  2019-01-24
#> 29       37  2016-04-06
#> 30       38  2015-04-27
#> 31       39  2019-01-25
#> 32       40  2019-09-19
#> 33       41  2015-05-22
#> 34       43  2018-05-22
#> 35       44  2015-06-11
#> 36       45  2019-05-07
#> 37       46  2019-12-23
#> 38       50  2017-09-21
#> 39       54  2019-12-17
#> 40       57  2020-01-29
#> 41       58  2019-04-25
#> 42       59  2018-01-30
#> 43       60  2018-04-19
#> 44       61  2019-04-25
#> 45       62  2017-05-09
#> 46       63  2018-10-10
#> 47       64  2018-12-03
#> 48       65  2019-11-06
#> 49       66  2019-10-29
#> 50       67  2019-10-23
#> 51       68  2019-11-27
#> 52       69  2019-06-27
#> 53       70  2019-04-26
#> 54       71  2008-02-27
#> 55       72  2019-09-10
#> 56       73  2019-06-27
#> 57       75  2020-01-14
#>                                                                  source
#> 1                                                        Doing Business
#> 2                                          World Development Indicators
#> 3                                       Worldwide Governance Indicators
#> 4                                     Subnational Malnutrition Database
#> 5                                         International Debt Statistics
#> 6                                         Africa Development Indicators
#> 7                                                  Education Statistics
#> 8                                                    Enterprise Surveys
#> 9                                                     Gender Statistics
#> 10                                              Global Economic Monitor
#> 11                           Health Nutrition and Population Statistics
#> 12                                       IDA Results Measurement System
#> 13                                         Millennium Development Goals
#> 14                                         Quarterly Public Sector Debt
#> 15                              Quarterly External Debt Statistics SDDS
#> 16                              Quarterly External Debt Statistics GDDS
#> 17                                                   Poverty and Equity
#> 18                                                                 Jobs
#> 19                                            Global Economic Prospects
#> 20                                           Global Financial Inclusion
#> 21  The Atlas of Social Protection: Indicators of Resilience and Equity
#> 22        Exporter Dynamics Database – Indicators at Country-Year Level
#> 23                          Country Policy and Institutional Assessment
#> 24                                         Global Financial Development
#> 25                                   G20 Financial Inclusion Indicators
#> 26                                     Global Partnership for Education
#> 27                                           Sustainable Energy for All
#> 28                                      Statistical Capacity Indicators
#> 29                                                       LAC Equity Lab
#> 30                                                  Subnational Poverty
#> 31        Health Nutrition and Population Statistics by Wealth Quintile
#> 32                                 Population estimates and projections
#> 33                 Country Partnership Strategy for India (FY2013 - 17)
#> 34                                                 Adjusted Net Savings
#> 35                       Readiness for Investment in Sustainable Energy
#> 36                  Indonesia Database for Policy and Economic Research
#> 37                                       Sustainable Development Goals 
#> 38                                               Subnational Population
#> 39                                              Joint External Debt Hub
#> 40                                                WDI Database Archives
#> 41                                            Universal Health Coverage
#> 42                                                      Wealth Accounts
#> 43                                                     Economic Fitness
#> 44                                              PPPs Regulatory Quality
#> 45                          International Comparison Program (ICP) 2011
#> 46                                                  Human Capital Index
#> 47                                     Worldwide Bureaucracy Indicators
#> 48                    Health Equity and Financial Protection Indicators
#> 49                                          Logistics Performance Index
#> 50                                                            PEFA 2011
#> 51                                                            PEFA 2016
#> 52            Global Financial Inclusion and Consumer Protection Survey
#> 53                                                   Economic Fitness 2
#> 54                          International Comparison Program (ICP) 2005
#> 55                                                            PEFA_Test
#> 56 Global Financial Inclusion and Consumer Protection Survey (Internal)
#> 57                        Environment, Social and Governance (ESG) Data
#>    sourceAbbr sourceDesc sourceURL dataAvail metadataAvail concepts
#> 1         DBS       <NA>      <NA>         Y             Y        3
#> 2         WDI       <NA>      <NA>         Y             Y        3
#> 3         WGI       <NA>      <NA>         Y             Y        3
#> 4         SNM       <NA>      <NA>         Y             Y        3
#> 5         IDS       <NA>      <NA>         Y             Y        3
#> 6         ADI       <NA>      <NA>         Y             Y        3
#> 7         EDS       <NA>      <NA>         Y             Y        3
#> 8         ESY       <NA>      <NA>         Y             Y        3
#> 9         GDS       <NA>      <NA>         Y             Y        3
#> 10        GEM       <NA>      <NA>         Y             Y        3
#> 11        HNP       <NA>      <NA>         Y             Y        3
#> 12        IDA       <NA>      <NA>         Y             Y        3
#> 13        MDG       <NA>      <NA>         Y             Y        3
#> 14        PSD       <NA>      <NA>         Y             Y        3
#> 15        QDS       <NA>      <NA>         Y             Y        3
#> 16        QDG       <NA>      <NA>         Y             Y        3
#> 17        POV       <NA>      <NA>         Y             Y        3
#> 18        JOB       <NA>      <NA>         Y             Y        3
#> 19        GEP       <NA>      <NA>         Y             N        3
#> 20        FDX       <NA>      <NA>         Y             Y        3
#> 21        GSP       <NA>      <NA>         Y             Y        3
#> 22        ED1       <NA>      <NA>         Y             N        3
#> 23        CPI       <NA>      <NA>         Y             Y        3
#> 24        GFD       <NA>      <NA>         Y             N        3
#> 25        G2F       <NA>      <NA>         Y             Y        3
#> 26        GPE       <NA>      <NA>         Y             Y        3
#> 27        SE4       <NA>      <NA>         Y             Y        3
#> 28        BBS       <NA>      <NA>         Y          <NA>        3
#> 29        LEL       <NA>      <NA>         Y             Y        3
#> 30        SNP       <NA>      <NA>         Y          <NA>        3
#> 31        HNQ       <NA>      <NA>         Y             Y        3
#> 32        HPP       <NA>      <NA>         Y             Y        3
#> 33        CPS       <NA>      <NA>         Y             N        3
#> 34        WAT       <NA>      <NA>         Y             N        3
#> 35        RIS       <NA>      <NA>         Y          <NA>        3
#> 36        IDD       <NA>      <NA>         Y          <NA>        3
#> 37        SDG       <NA>      <NA>         Y          <NA>        3
#> 38        SNT       <NA>      <NA>         Y          <NA>        3
#> 39        JED       <NA>      <NA>         Y          <NA>        3
#> 40        WDA       <NA>      <NA>         Y             Y        4
#> 41        UHC       <NA>      <NA>         Y             Y        3
#> 42        WAC       <NA>      <NA>         Y             Y        3
#> 43        EFT       <NA>      <NA>         Y             Y        3
#> 44        PRQ       <NA>      <NA>         Y             Y        3
#> 45        ICP       <NA>      <NA>         N             Y        4
#> 46        HCI       <NA>      <NA>         Y             Y        3
#> 47        WBI       <NA>      <NA>         Y             N        3
#> 48        HPI       <NA>      <NA>         Y             Y        3
#> 49        LPI       <NA>      <NA>         Y             Y        3
#> 50        PF1       <NA>      <NA>         Y             Y        4
#> 51        PF6       <NA>      <NA>         Y             Y        4
#> 52        RFA       <NA>      <NA>         Y             Y        3
#> 53        EF2       <NA>      <NA>         Y             Y        3
#> 54        IC5       <NA>      <NA>         Y             N        4
#> 55        PFT       <NA>      <NA>         Y             Y        4
#> 56        RFI       <NA>      <NA>         Y             Y        3
#> 57        ESG       <NA>      <NA>         Y             Y        3

wbsources(lang = "es")

#>    sourceID lastUpdated
#> 1         1  2019-10-23
#> 2         2  2019-12-20
#> 3         3  2019-11-07
#> 4         5  2016-03-21
#> 5         6  2019-12-02
#> 6        11  2013-02-22
#> 7        12  2019-12-10
#> 8        13  2017-01-03
#> 9        14  2019-12-20
#> 10       15  2019-09-13
#> 11       16  2019-12-20
#> 12       18  2019-12-23
#> 13       19  2018-09-19
#> 14       20  2020-01-27
#> 15       22  2020-01-31
#> 16       23  2020-01-31
#> 17       24  2019-09-24
#> 18       25  2019-12-20
#> 19       27  2020-01-08
#> 20       28  2018-10-15
#> 21       29  2019-05-02
#> 22       30  2016-03-31
#> 23       31  2019-08-22
#> 24       32  2019-10-30
#> 25       33  2019-02-27
#> 26       34  2013-04-12
#> 27       35  2018-06-30
#> 28       36  2019-01-24
#> 29       37  2016-04-06
#> 30       38  2015-04-27
#> 31       39  2019-01-25
#> 32       40  2019-09-19
#> 33       41  2015-05-22
#> 34       43  2018-05-22
#> 35       44  2015-06-11
#> 36       45  2019-05-07
#> 37       46  2019-12-23
#> 38       50  2017-09-21
#> 39       54  2019-12-17
#> 40       57  2020-01-29
#> 41       58  2019-04-25
#> 42       59  2018-01-30
#> 43       60  2018-04-19
#> 44       61  2019-04-25
#> 45       62  2017-05-09
#> 46       63  2018-10-10
#> 47       64  2018-12-03
#> 48       65  2019-11-06
#> 49       66  2019-10-29
#> 50       67  2019-10-23
#> 51       68  2019-11-27
#> 52       69  2019-06-27
#> 53       70  2019-04-26
#> 54       71  2008-02-27
#> 55       72  2019-09-10
#> 56       73  2019-06-27
#> 57       75  2020-01-14
#>                                                                     source
#> 1                                                           Doing Business
#> 2                                       Indicadores del desarrollo mundial
#> 3                                            Indicadores de gobernabilidad
#> 4                                                                     <NA>
#> 5                                                                     <NA>
#> 6                                      Indicadores de desarrollo de África
#> 7                                             Estadísticas sobre educación
#> 8                                                    Encuestas de empresas
#> 9                                                Estadísticas sobre género
#> 10          Monitor de la Economía Mundial (GEM, por sus siglas en inglés)
#> 11                         Estadísticas sobre salud, nutrición y población
#> 12 Asociación Internacional de Fomento – Sistema de medición de resultados
#> 13                                     Objetivos de desarrollo del milenio
#> 14                                                                    <NA>
#> 15                                                                    <NA>
#> 16                                                                    <NA>
#> 17                                                                    <NA>
#> 18                                                                    <NA>
#> 19                                                  GEP Economic Prospects
#> 20                                                                    <NA>
#> 21                                                                    <NA>
#> 22                                                                    <NA>
#> 23                                                                    <NA>
#> 24                                                                    <NA>
#> 25                                                                    <NA>
#> 26                                                                    <NA>
#> 27                                                                    <NA>
#> 28                                                                    <NA>
#> 29                                                                    <NA>
#> 30                                                                    <NA>
#> 31                                                                    <NA>
#> 32                                                                    <NA>
#> 33                                                                    <NA>
#> 34                                                                    <NA>
#> 35                                                                    <NA>
#> 36                                                                    <NA>
#> 37                                                                    <NA>
#> 38                                                                    <NA>
#> 39                                                                    <NA>
#> 40                                                                    <NA>
#> 41                                                                    <NA>
#> 42                                                                    <NA>
#> 43                                                                    <NA>
#> 44                                                                    <NA>
#> 45                                                                    <NA>
#> 46                                                                    <NA>
#> 47                                                                    <NA>
#> 48                                                                    <NA>
#> 49                                                                    <NA>
#> 50                                                                    <NA>
#> 51                                                                    <NA>
#> 52                                                                    <NA>
#> 53                                                                    <NA>
#> 54                                                                    <NA>
#> 55                                                                    <NA>
#> 56                                                                    <NA>
#> 57                                                                    <NA>
#>    sourceAbbr sourceDesc sourceURL dataAvail metadataAvail concepts
#> 1         DBS       <NA>      <NA>         Y             Y        3
#> 2         WDI       <NA>      <NA>         Y             Y        3
#> 3         WGI       <NA>      <NA>         Y             Y        3
#> 4         SNM       <NA>      <NA>         Y             Y        3
#> 5         IDS       <NA>      <NA>         Y             Y        3
#> 6         ADI       <NA>      <NA>         Y             Y        3
#> 7         EDS       <NA>      <NA>         Y             Y        3
#> 8         ESY       <NA>      <NA>         Y             Y        3
#> 9         GDS       <NA>      <NA>         Y             Y        3
#> 10        GEM       <NA>      <NA>         Y             Y        3
#> 11        HNP       <NA>      <NA>         Y             Y        3
#> 12        IDA       <NA>      <NA>         Y             Y        3
#> 13        MDG       <NA>      <NA>         Y             Y        3
#> 14        PSD       <NA>      <NA>         Y             Y        3
#> 15        QDS       <NA>      <NA>         Y             Y        3
#> 16        QDG       <NA>      <NA>         Y             Y        3
#> 17        POV       <NA>      <NA>         Y             Y        3
#> 18        JOB       <NA>      <NA>         Y             Y        3
#> 19        GEP       <NA>      <NA>         Y             N        3
#> 20        FDX       <NA>      <NA>         Y             Y        3
#> 21        GSP       <NA>      <NA>         Y             Y        3
#> 22        ED1       <NA>      <NA>         Y             N        3
#> 23        CPI       <NA>      <NA>         Y             Y        3
#> 24        GFD       <NA>      <NA>         Y             N        3
#> 25        G2F       <NA>      <NA>         Y             Y        3
#> 26        GPE       <NA>      <NA>         Y             Y        3
#> 27        SE4       <NA>      <NA>         Y             Y        3
#> 28        BBS       <NA>      <NA>         Y          <NA>        3
#> 29        LEL       <NA>      <NA>         Y             Y        3
#> 30        SNP       <NA>      <NA>         Y          <NA>        3
#> 31        HNQ       <NA>      <NA>         Y             Y        3
#> 32        HPP       <NA>      <NA>         Y             Y        3
#> 33        CPS       <NA>      <NA>         Y             N        3
#> 34        WAT       <NA>      <NA>         Y             N        3
#> 35        RIS       <NA>      <NA>         Y          <NA>        3
#> 36        IDD       <NA>      <NA>         Y          <NA>        3
#> 37        SDG       <NA>      <NA>         Y          <NA>        3
#> 38        SNT       <NA>      <NA>         Y          <NA>        3
#> 39        JED       <NA>      <NA>         Y          <NA>        3
#> 40        WDA       <NA>      <NA>         Y             Y        4
#> 41        UHC       <NA>      <NA>         Y             Y        3
#> 42        WAC       <NA>      <NA>         Y             Y        3
#> 43        EFT       <NA>      <NA>         Y             Y        3
#> 44        PRQ       <NA>      <NA>         Y             Y        3
#> 45        ICP       <NA>      <NA>         N             Y        4
#> 46        HCI       <NA>      <NA>         Y             Y        3
#> 47        WBI       <NA>      <NA>         Y             N        3
#> 48        HPI       <NA>      <NA>         Y             Y        3
#> 49        LPI       <NA>      <NA>         Y             Y        3
#> 50        PF1       <NA>      <NA>         Y             Y        4
#> 51        PF6       <NA>      <NA>         Y             Y        4
#> 52        RFA       <NA>      <NA>         Y             Y        3
#> 53        EF2       <NA>      <NA>         Y             Y        3
#> 54        IC5       <NA>      <NA>         Y             N        4
#> 55        PFT       <NA>      <NA>         Y             Y        4
#> 56        RFI       <NA>      <NA>         Y             Y        3
#> 57        ESG       <NA>      <NA>         Y             Y        3