Download updated indicator topic information from World Bank API

Download updated information on available indicator topics from the World Bank API

wbtopics(lang = c("en", "es", "fr", "ar", "zh"))

Arguments

lang
Language in which to return the results. If lang is unspecified, english is the default.

Value

A data frame of available indicator topics with related information

Note

Not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA. For an enumeration of supported languages by data source please see wbdatacatalog. The options for lang are:

en: English

es: Spanish

fr: French

ar: Arabic

zh: Mandarin

Examples

wbtopics()

#>    topicID                           topic
#> 1        1 Agriculture & Rural Development
#> 2        2               Aid Effectiveness
#> 3        3                Economy & Growth
#> 4        4                       Education
#> 5        5                 Energy & Mining
#> 6        6                     Environment
#> 7        7                Financial Sector
#> 8        8                          Health
#> 9        9                  Infrastructure
#> 10      10       Social Protection & Labor
#> 11      11                         Poverty
#> 12      12                  Private Sector
#> 13      13                   Public Sector
#> 14      14            Science & Technology
#> 15      15              Social Development
#> 16      16               Urban Development
#> 17      17                          Gender
#> 18      18     Millenium development goals
#> 19      19                  Climate Change
#> 20      20                   External Debt
#> 21      21                           Trade
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    topicDesc
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              For the 70 percent of the world's poor who live in rural areas, agriculture is the main source of income and employment. But depletion and degradation of land and water pose serious challenges to producing enough food and other agricultural products to sustain livelihoods here and meet the needs of urban populations. Data presented here include measures of agricultural inputs, outputs, and productivity compiled by the UN's Food and Agriculture Organization.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Aid effectiveness is the impact that aid has in reducing poverty and inequality, increasing growth, building capacity, and accelerating achievement of the Millennium Development Goals set by the international community. Indicators here cover aid received as well as progress in reducing poverty and improving education, health, and other measures of human welfare.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                              Economic growth is central to economic development. When national income grows, real people benefit. While there is no known formula for stimulating economic growth, data can help policy-makers better understand their countries' economic situations and guide any work toward improvement. Data here covers measures of economic growth, such as gross domestic product (GDP) and gross national income (GNI). It also includes indicators representing factors known to be relevant to economic growth, such as capital stock, employment, investment, savings, consumption, government spending, imports, and exports.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Education is one of the most powerful instruments for reducing poverty and inequality and lays a foundation for sustained economic growth. The World Bank compiles data on education inputs, participation, efficiency, and outcomes. Data on education are compiled by the United Nations Educational, Scientific, and Cultural Organization (UNESCO) Institute for Statistics from official responses to surveys and from reports provided by education authorities in each country.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 The world economy needs ever-increasing amounts of energy to sustain economic growth, raise living standards, and reduce poverty. But today's trends in energy use are not sustainable. As the world's population grows and economies become more industrialized, nonrenewable energy sources will become scarcer and more costly. Data here on energy production, use, dependency, and efficiency are compiled by the World Bank from the International Energy Agency and the Carbon Dioxide Information Analysis Center.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                       Natural and man-made environmental resources – fresh water, clean air, forests, grasslands, marine resources, and agro-ecosystems – provide sustenance and a foundation for social and economic development.  The need to safeguard these resources crosses all borders.  Today, the World Bank is one of the key promoters and financiers of environmental upgrading in the developing world. Data here cover forests, biodiversity, emissions, and pollution. Other indicators relevant to the environment are found under data pages for Agriculture & Rural Development, Energy & Mining, Infrastructure, and Urban Development.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                             An economy's financial markets are critical to its overall development. Banking systems and stock markets enhance growth, the main factor in poverty reduction. Strong financial systems provide reliable and accessible information that lowers transaction costs, which in turn bolsters resource allocation and economic growth. Indicators here include the size and liquidity of stock markets; the accessibility, stability, and efficiency of financial systems; and international migration and workers\\ remittances, which affect growth and social welfare in both sending and receiving countries.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                Improving health is central to the Millennium Development Goals, and the public sector is the main provider of health care in developing countries. To reduce inequities, many countries have emphasized primary health care, including immunization, sanitation, access to safe drinking water, and safe motherhood initiatives.  Data here cover health systems, disease prevention, reproductive health, nutrition, and population dynamics. Data are from the United Nations Population Division, World Health Organization, United Nations Children's Fund, the Joint United Nations Programme on HIV/AIDS, and various other sources.
#> 9                                                                                                                                                                                                                                                                                                                                                                                       Infrastructure helps determine the success of manufacturing and agricultural activities. Investments in water, sanitation, energy, housing, and transport also improve lives and help reduce poverty. And new information and communication technologies promote growth, improve delivery of health and other services, expand the reach of education, and support social and cultural advances. Data here are compiled from such sources as the International Road Federation, Containerisation International, the International Civil Aviation Organization, the International Energy Association, and the International Telecommunications Union.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                             The supply of labor available in an economy includes people who are employed, those who are unemployed but seeking work, and first-time job-seekers. Not everyone who works is included: unpaid workers, family workers, and students are often omitted, while some countries do not count members of the armed forces. Data on labor and employment are compiled by the International Labour Organization (ILO) from labor force surveys, censuses, establishment censuses and surveys, and administrative records such as employment exchange registers and unemployment insurance schemes.
#> 11                                                                                                                                                                                                                                                                                                                                                                  For countries with an active poverty monitoring program, the World Bank—in collaboration with national institutions, other development agencies, and civil society—regularly conducts analytical work to assess the extent and causes of poverty and inequality, examine the impact of growth and public policy, and review household survey data and measurement methods.  Data here includes poverty and inequality measures generated from analytical reports, from national poverty monitoring programs, and from the World Bank’s Development Research Group which has been producing internationally comparable and global poverty estimates and lines since 1990.
#> 12                                                                                                                                                                                                                                                                                                                                                      Private markets drive economic growth, tapping initiative and investment to create productive jobs and raise incomes. Trade is also a driver of economic growth as it integrates developing countries into the world economy and generates benefits for their people.  Data on the private sector and trade are from the World Bank Group's Private Participation in Infrastructure Project Database, Enterprise Surveys, and Doing Business Indicators, as well as from the International Monetary Fund's Balance of Payments database and International Financial Statistics, the UN Commission on Trade and Development, the World Trade Organization, and various other sources.
#> 13                                                                                                                                                                                                                                                                                                                                                                                          Effective governments improve people's standard of living by ensuring access to essential services –  health, education, water and sanitation, electricity, transport – and the opportunity to live and work in peace and security. Data here includes World Bank staff assessments of country performance in economic management, structural policies, policies for social inclusion and equity, and public sector management and institutions for the poorest countries. Also included are indicators on revenues and expenses from the International Monetary Fund's Government Finance Statistics, and on tax policies from various sources.
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Technological innovation, often fueled by governments, drives industrial growth and helps raise living standards. Data here aims to shed light on countries technology base: research and development, scientific and technical journal articles, high-technology exports, royalty and license fees, and patents and trademarks. Sources include the UNESCO Institute for Statistics, the U.S. National Science Board, the UN Statistics Division, the International Monetary Fund, and the World Intellectual Property Organization.
#> 15                                                                                                                                                                                                                                                                                           Data here cover child labor, gender issues, refugees, and asylum seekers. Children in many countries work long hours, often combining studying with work for pay. The data on their paid work are from household surveys conducted by the International Labour Organization (ILO), the United Nations Children's Fund (UNICEF), the World Bank, and national statistical offices. Gender disparities are measured using a compilation of data on key topics such as education, health, labor force participation, and political participation.  Data on refugees are from the United Nations High Commissioner for Refugees complemented by statistics on Palestinian refugees under the mandate of the United Nations Relief and Works Agency.
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                             Cities can be tremendously efficient. It is easier to provide water and sanitation to people living closer together, while access to health, education, and other social and cultural services is also much more readily available. However, as cities grow, the cost of meeting basic needs increases, as does the strain on the environment and natural resources. Data on urbanization, traffic and congestion, and air pollution are from the United Nations Population Division, World Health Organization, International Road Federation, World Resources Institute, and other sources.
#> 17                                                                                                                    Gender equality is a core development objective in its own right. It is also smart development policy and sound business practice. It is integral to economic growth, business growth and good development outcomes. Gender equality can boost productivity, enhance prospects for the next generation, build resilience, and make institutions more representative and effective.  In December 2015, the World Bank Group Board discussed our new Gender Equality Strategy 2016-2023, which aims to address persistent gaps and proposed a sharpened focus on more and better gender data. The Bank Group is continually scaling up commitments and expanding partnerships to fill significant gaps in gender data. The database hosts the latest sex-disaggregated data and gender statistics covering demography, education, health, access to economic opportunities, public life and decision-making, and agency.
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      <NA>
#> 19 Climate change is expected to hit developing countries the hardest. Its effects—higher temperatures, changes in precipitation patterns, rising sea levels, and more frequent weather-related disasters—pose risks for agriculture, food, and water supplies. At stake are recent gains in the fight against poverty, hunger and disease, and the lives and livelihoods of billions of people in developing countries. Addressing climate change requires unprecedented global cooperation across borders. The World Bank Group is helping support developing countries and contributing to a global solution, while tailoring our approach to the differing needs of developing country partners.  Data here cover climate systems, exposure to climate impacts, resilience, greenhouse gas emissions, and energy use.  Other indicators relevant to climate change are found under other data pages, particularly Environment, Agriculture & Rural Development, Energy & Mining, Health, Infrastructure, Poverty, and Urban Development.
#> 20                                                                                                                                                                                                                                                                                                           Debt statistics provide a detailed picture of debt stocks and flows of developing countries. Data presented as part of the Quarterly External Debt Statistics takes a closer look at the external debt of high-income countries and emerging markets to enable a more complete understanding of global financial flows. The Quarterly Public Sector Debt database provides further data on public sector valuation methods, debt instruments, and clearly defined tiers of debt for central, state and local government, as well as extra-budgetary agencies and funds. Data are gathered from national statistical organizations and central banks as well as by various major multilateral institutions and World Bank staff.
#> 21                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Trade is a key means to fight poverty and achieve the Millennium Development Goals, specifically by improving developing country access to markets, and supporting a rules based, predictable trading system. In cooperation with other international development partners, the World Bank launched the Transparency in Trade Initiative to provide free and easy access to data on country-specific trade policies.

wbtopics(lang = "es")

#>    topicID                               topic
#> 1        1      Agricultura y desarrollo rural
#> 2        2                Eficacia de la ayuda
#> 3        3              Economía y crecimiento
#> 4        4                           Educación
#> 5        5                   Energía y minería
#> 6        6                      Medio ambiente
#> 7        7                   Sector financiero
#> 8        8                               Salud
#> 9        9                     Infraestructura
#> 10      10         Protección social y trabajo
#> 11      11                             Pobreza
#> 12      12                      Sector privado
#> 13      13                      Sector público
#> 14      14                Ciencia y tecnología
#> 15      15                   Desarrollo social
#> 16      16                   Desarrollo urbano
#> 17      17                              Género
#> 18      18 Objetivos de desarrollo del milenio
#> 19      19                    Cambio climático
#> 20      20                       Deuda externa
#> 21      21                            Comercio
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 topicDesc
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Para el 70% de los pobres del mundo que viven en zonas rurales, la agricultura es la principal fuente de ingresos y de trabajo. Pero el agotamiento y la degradación de la tierra y del agua afectan gravemente la capacidad de cultivar alimentos y otros productos necesarios para sustentar los medios de vida en estas zonas y satisfacer las necesidades de la población urbana. Los datos presentados contienen medidas de los insumos, productos y productividad agrícola compiladas por la Organización de las Naciones Unidas para la Agricultura y la Alimentación (FAO).
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Esta página se refiere al impacto de la asistencia en la reducción de la pobreza y la desigualdad, el fomento del crecimiento, la formación de capacidades y el aceleramiento de la consecución de los objetivos de desarrollo del milenio (ODM) fijados por la comunidad internacional. Estos indicadores cubren la ayuda recibida y los avances en la lucha contra la pobreza y el mejoramiento de la educación, la salud y otras medidas del bienestar humano.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 El crecimiento de la economía constituye un aspecto central del desarrollo económico.  Cuando el ingreso de los países aumenta, se benefician las personas. Si bien no existe una fórmula conocida para estimular el crecimiento económico, los datos pueden ayudar a los encargados de formular políticas a entender mejor la situación económica de sus países y orientar cualquier esfuerzo destinado a mejorarla. Los datos entregados aquí abarcan indicadores del crecimiento económico, como el producto interno bruto (PIB) y el ingreso nacional bruto (INB).  También se ofrecen otros indicadores que son relevantes para el crecimiento económico, entre ellos:  capital, empleo, inversión, ahorro, consumo, gasto público, importaciones y exportaciones.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Se trata de uno de los instrumentos más poderosos para reducir la pobreza y la desigualdad y sienta las bases para un crecimiento económico sostenido. El Banco Mundial compila datos sobre los insumos, participación, eficiencia y resultados del sector. La información sobre el tema es recolectada por el Instituto de Estadística de la Organización de las Naciones Unidas para la Educación, la Ciencia y la Cultura (UNESCO) a partir de respuestas oficiales a encuestas y de informes provistos por las autoridades sectoriales en cada país.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                La economía mundial necesita cada vez más energía para mantener el crecimiento económico, mejorar los niveles de vida y reducir la pobreza. Pero las actuales tendencias en su uso no son sostenibles. A medida que aumenta la población mundial y las economías se industrializan, las fuentes de energía no renovable se tornarán más escasas y costosas. La Agencia Internacional de la Energía (AIE) y el Centro de Análisis de la información sobre el dióxido de carbono de los Estados Unidos de América son las fuentes de estos datos sobre producción, uso, dependencia y eficiencia energética que recopila el Banco Mundial.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Los recursos naturales y los creados por el hombre –agua limpia, aire puro, bosques, praderas, recursos marinos y ecosistemas agrícolas– proporcionan sustento y una base para el desarrollo social y económico. La necesidad de protegerlos trasciende las fronteras. Hoy en día, el Banco Mundial es uno de los principales promotores y financistas de los programas de mejoramiento ambiental en el mundo en desarrollo. Estos datos se refieren a bosques, biodiversidad, emisiones y contaminación. Otros indicadores importantes sobre el medio ambiente se encuentran en las páginas sobre agricultura y desarrollo rural, energía y minería, infraestructura y desarrollo urbano.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Los mercados financieros son fundamentales para la evolución general de una economía. Los sistemas bancarios y los mercados accionarios estimulan el crecimiento, el factor que más incide en la reducción de la pobreza. Si los sistemas financieros son sólidos, se cuenta con información confiable y accesible que reduce los costos de transacción. Esto a la vez mejora la asignación de recursos y el crecimiento económico. Estos indicadores abarcan el tamaño y la liquidez de los mercados accionarios; la accesibilidad, estabilidad y eficiencia de los sistemas financieros; y la migración internacional y las remesas de los trabajadores, aspectos que inciden en el crecimiento y el bienestar social, tanto en los países de origen como de destino.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Mejorar la salud es un aspecto central de los ODM y el sector público es el principal proveedor de estos servicios en los países en desarrollo. Para reducir las inequidades, muchas naciones han puesto énfasis en la atención primaria, campo que incluye vacunación, saneamiento, acceso a agua potable e iniciativas de maternidad sin riesgos. Estos datos incluyen sistemas de salud, calidad y cobertura de la prevención, salud reproductiva, nutrición y dinámica demográfica. La información proviene de la División de Población de las Naciones Unidas (ONU), la Organización Mundial de la Salud (OMS), el Fondo de las Naciones Unidas para la Infancia (UNICEF), el Programa Conjunto de las Naciones Unidas contra el VIH/SIDA y diversas otras fuentes.
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             La infraestructura determina el éxito de las actividades manufactureras y agrícolas. Las inversiones en agua, saneamiento, energía, vivienda y transporte también mejoran la calidad de vida y ayudan a reducir la pobreza. Además, las nuevas tecnologías de la información y las comunicaciones promueven el crecimiento, mejoran la prestación de servicios de salud y otros, amplían el alcance de la educación y apoyan los avances sociales y culturales. Estos datos se recopilan de fuentes como la Federación Internacional de Carreteras, la Organización de Aviación Civil Internacional, la Asociación Internacional de Energía y la Unión Internacional de Telecomunicaciones.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       La oferta de mano de obra disponible en una economía incluye a personas que tienen empleo y a quienes no lo tienen, pero están buscando, así como también a los que intentan encontrar trabajo por primera vez. No todos los que trabajan son considerados; a menudo se omite a quienes laboran sin salario, a los trabajadores de familia no remunerados y a los estudiantes, mientras que algunos países no consideran a los miembros de las Fuerzas Armadas. La Organización Internacional del Trabajo (OIT) compila datos provenientes de encuestas y censos sobre mano de obra, empleo y establecimientos, registros administrativos de cambios de trabajo y planes de seguros de cesantía.
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                Para los países con un programa activo de seguimiento de la pobreza, el Banco Mundial —en colaboración con instituciones nacionales, otros organismos de desarrollo y la sociedad civil— realiza periódicamente tareas de análisis para evaluar el alcance y las causas de la pobreza y la desigualdad, examinar el impacto del crecimiento y las políticas públicas, y revisar los datos de encuestas de hogares y los métodos de medición. Los presentes datos incluyen medidas de desigualdad y pobreza generadas a partir de informes analíticos, programas nacionales de vigilancia de la pobreza y el Grupo de Investigación sobre Desarrollo del Banco Mundial, que ha estado produciendo estimaciones y líneas de pobreza mundial comparables a nivel internacional desde 1990.
#> 12                                                                                                                                                                                                                                                                                                                                                           Los mercados privados estimulan el crecimiento económico gracias a que aprovechan la iniciativa y las inversiones para crear empleos productivos y aumentar los ingresos. El comercio también es un motor del crecimiento en la medida en que integra a los países en desarrollo al concierto de la economía mundial y genera beneficios para sus habitantes. Los datos sobre el sector privado y el comercio provienen de la base de datos del Proyecto sobre Participación Privada en Obras de Infraestructura, encuestas a empresas e indicadores de Doing Business del Grupo del Banco Mundial, de la base de datos de la balanza de pagos y las Estadísticas Financieras Internacionales del FMI, de la Conferencia de las Naciones Unidas sobre Comercio y Desarrollo (UNCTAD, por sus siglas en inglés), de la Organización Mundial del Comercio (OMC) y de muchas otras fuentes más.
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Los Gobiernos eficientes mejoran el nivel de vida de sus pueblos garantizando el acceso a servicios esenciales –salud, educación, agua y saneamiento, electricidad y transporte– y la oportunidad de vivir y trabajar en condiciones de paz y seguridad. Estos datos contienen evaluaciones realizadas por el personal del Banco sobre el desempeño de cada país en gestión económica, políticas estructurales, medidas de inclusión social y equidad, gestión del sector público e instituciones de los países más pobres. También es incluyen indicadores sobre ingresos y gastos de las Estadísticas de las Finanzas Públicas del FMI y sobre políticas tributarias de muchas fuentes.
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             La innovación tecnológica, normalmente estimulada por los Gobiernos, promueve el crecimiento industrial y ayuda a mejorar los niveles de vida de la población. Estos datos pretenden arrojar luz sobre la base tecnológica de los países e incluyen: investigación y desarrollo, artículos de publicaciones científicas y técnicas, exportaciones de alta tecnología, regalías y derechos de licencias y patentes y marcas comerciales. Las fuentes incluyen al Instituto de Estadística de la UNESCO, la Fundación Nacional de las Ciencias de Estados Unidos, la División de Estadística de la ONU, el FMI y la Organización Mundial de la Propiedad Intelectual (OMPI).
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                              Estos datos cubren trabajo infantil, temas de género, refugiados y personas que piden asilo. En muchos países, los niños laboran prolongadas jornadas y a menudo combinan estudios con trabajo remunerado. Los datos sobre este último tema provienen de encuestas por hogares realizadas por la OIT, el UNICEF, el Banco Mundial e instituciones nacionales de estadísticas. Las disparidades de género se miden usando una compilación de datos sobre temas clave como educación, salud y participación política y en la fuerza laboral. La información sobre refugiados proviene del Alto Comisionado de las Naciones Unidas para los Refugiados (ACNUR), complementada por estadísticas sobre refugiados palestinos bajo el mandato del Organismo de Obras Públicas y Socorro de las Naciones Unidas.
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Las ciudades pueden ser extremadamente eficientes. Es más fácil suministrar servicios de agua y saneamiento cuando las personas viven más concentradas en el espacio. También se facilita el acceso a salud, educación y otros servicios sociales y culturales. Sin embargo, a medida que las urbes crecen, aumenta el costo de satisfacer las necesidades básicas y se intensifica la presión sobre el medio ambiente y los recursos naturales. Los datos sobre urbanización, tráfico y congestión y contaminación del aire provienen de la División de Población de las Naciones Unidas, la OMS, la Federación Internacional de Carreteras, el Instituto de Recursos Mundiales y otras múltiples fuentes.
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                                   Empoderar a la mujer y promover la igualdad de género es clave para lograr un desarrollo sostenible. Una mayor equidad de género puede hacer más eficiente la economía y mejorar otros resultados en materia de desarrollo al eliminar barreras que impiden a las mujeres tener el mismo acceso que los hombres a las dotaciones de personal, derechos y oportunidades económicas. Dar a las mujeres acceso a iguales oportunidades les permite tener un rol social y económico más relevante y así se puede avanzar en la formulación de políticas más inclusivas. Además, esto lleva a aumentar las inversiones en la educación, salud y bienestar general de sus hijos.  Los datos presentados aquí incluyen información sobre demografía, educación, salud, fuerza laboral y empleo, y participación política.  
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   <NA>
#> 19 Se prevé que el cambio climático afectará más gravemente a los países en desarrollo. Sus efectos -altas temperaturas, cambios en el régimen de precipitaciones, aumento del nivel del mar y desastres más frecuentes relacionados con el clima- representan riesgos para la agricultura, los alimentos y el suministro de agua. Están en juego los recientes avances en la lucha contra la pobreza, el hambre y las enfermedades y la vida y el sustento de miles de millones de personas de estas naciones. Abordar el cambio climático requiere una cooperación mundial sin precedentes. El Grupo del Banco Mundial está ayudando a apoyar a los países en desarrollo y contribuyendo a una solución global mediante la adaptación de su estrategia a las diferentes necesidades de dichos países asociados. Los datos que se presentan aquí cubren los sistemas climáticos, la exposición a los impactos del clima, la adaptación, las emisiones de gases de efecto invernadero y el consumo de energía. Otros indicadores relevantes para el cambio climático se encuentran en otras páginas de datos, en particular sobre Medio Ambiente, Agricultura y Desarrollo Rural, Energía y Minería, Salud, Infraestructura, Pobreza y Desarrollo Urbano.
#> 20                                                                                                                                                                                                                                                                                                                                                  Las Estadísticas de la deuda ofrecen un panorama detallado del saldo y los flujos de deuda de los países en desarrollo. En los datos presentados, como parte de las Estadísticas trimestrales de la deuda externa, se analiza con mayor detenimiento la deuda externa de los países de ingreso alto y de los mercados emergentes para permitir una comprensión más completa de los flujos financieros mundiales. La base de datos de la Deuda trimestral del sector público entrega mayor información sobre métodos de valoración del sector público, instrumentos de deuda, y niveles de deuda claramente definidos para el gobierno central, estatal y local, así como acerca de organismos y fondos extrapresupuestarios. Los datos son obtenidos tanto de organismos nacionales de estadística y bancos centrales como de importantes instituciones multilaterales y personal del Banco Mundial. 
#> 21                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              El comercio es un medio clave para luchar contra la pobreza y alcanzar los objetivos de desarrollo del milenio (ODM), particularmente porque facilita el acceso de los países en desarrollo a los mercados y favorece un sistema de intercambio previsible y basado en reglas. Con la colaboración de otros asociados en el desarrollo, el Banco Mundial lanzó la iniciativa Transparencia en el Comercio para dar acceso fácil y gratuito a datos sobre políticas comerciales específicas de los países.