Download an updated list of country, indicator, and source information

Download an updated list of information regarding countries, indicators, sources, regions, indicator topics, lending types, income levels, and supported languages from the World Bank API

wb_cache(lang)

Arguments

lang
Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Value

A list containing the following items:

countries: The result of calling wb_countries()

indicators: The result of calling wb_indicators()

sources: The result of calling wb_sources()

topics: The result of calling wb_topics()

regions: The result of calling wb_regions()

income_levels: The result of calling wb_income_levels()

lending_types: The result of calling wb_lending_types()

languages: The result of calling wb_languages()

Note

Not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA. For an enumeration of supported languages by data source please see wb_languages()

Saving this return and using it has the cache parameter in wb_data() and wb_search() replaces the default cached version wb_cachelist that comes with the package itself

Examples

# the default is english. To specify another language use argument lang

# spanish
es_cache <- wb_cache(lang = "es")

# korean
ko_cache <- wb_cache(lang = "ko")

# possible values of lang are in the iso2 column in the languages data frame
example_iso2 <- wb_languages()$iso2[1]
wb_cache(lang = example_iso2)