Download an updated list of country, indicator, and source information

Download an updated list of information regarding countries, indicators, sources, regions, indicator topics, lending types, income levels, and supported languages from the World Bank API

wb_cache(lang)

Arguments

lang
Language in which to return the results. If lang is unspecified, english is the default. For supported languages see wb_languages(). Possible values of lang are in the iso2 column. A note of warning, not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA.

Value

A list containing the following items:

countries: The result of calling wb_countries()

indicators: The result of calling wb_indicators()

sources: The result of calling wb_sources()

topics: The result of calling wb_topics()

regions: The result of calling wb_regions()

income_levels: The result of calling wb_income_levels()

lending_types: The result of calling wb_lending_types()

languages: The result of calling wb_languages()

Note

Not all data returns have support for langauges other than english. If the specific return does not support your requested language by default it will return NA. For an enumeration of supported languages by data source please see wb_languages()

Saving this return and using it has the cache parameter in wb_data() and wb_search() replaces the default cached version wb_cachelist that comes with the package itself

Examples

# the default is english. To specify another language use argument lang

# spanish
wb_cache(lang = "es")

#> $countries
#> # A tibble: 304 x 18
#>    iso3c iso2c country capital_city longitude latitude region_iso3c region_iso2c
#>    <chr> <chr> <chr>   <chr>            <dbl>    <dbl> <chr>        <chr>       
#>  1 ABW   AW    Aruba   Oranjestad      -70.0     12.5  LCN          ZJ          
#>  2 AFG   AF    Afgani~ Kabul            69.2     34.5  SAS          8S          
#>  3 AFR   A9    NA      NA               NA       NA    NA           NA          
#>  4 AGO   AO    Angola  Luanda           13.2     -8.81 SSF          ZG          
#>  5 ALB   AL    Albania Tirana           19.8     41.3  ECS          Z7          
#>  6 AND   AD    Andorra Andorra la ~      1.52    42.5  ECS          Z7          
#>  7 ANR   L5    Región~ NA               NA       NA    NA           NA          
#>  8 ARB   1A    El mun~ NA               NA       NA    NA           NA          
#>  9 ARE   AE    Emirat~ Abu Dabi         54.4     24.5  MEA          ZQ          
#> 10 ARG   AR    Argent~ Buenos Aires    -58.4    -34.6  LCN          ZJ          
#> # ... with 294 more rows, and 10 more variables: region <chr>,
#> #   admin_region_iso3c <chr>, admin_region_iso2c <chr>, admin_region <chr>,
#> #   income_level_iso3c <chr>, income_level_iso2c <chr>, income_level <chr>,
#> #   lending_type_iso3c <chr>, lending_type_iso2c <chr>, lending_type <chr>
#> 
#> $indicators
#> # A tibble: 16,496 x 8
#>    indicator_id indicator unit  indicator_desc source_org topics source_id
#>    <chr>        <chr>     <lgl> <chr>          <chr>      <list>     <dbl>
#>  1 1.0.HCount.~ Tasa de ~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  2 1.0.HCount.~ Tasa de ~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  3 1.0.HCount.~ Tasa de ~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  4 1.0.HCount.~ Tasa Ofi~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  5 1.0.HCount.~ Tasa de ~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  6 1.0.HCount.~ Tasa de ~ NA    Tasa de Incid~ Tabulacio~ <df[,~        37
#>  7 1.0.PGap.1.~ Brecha d~ NA    La Brecha de ~ Tabulacio~ <df[,~        37
#>  8 1.0.PGap.2.~ Brecha d~ NA    La Brecha de ~ Tabulacio~ <df[,~        37
#>  9 1.0.PGap.Po~ Brecha d~ NA    La Brecha de ~ Tabulacio~ <df[,~        37
#> 10 1.0.PSev.1.~ Severida~ NA    El índice de ~ Tabulacio~ <df[,~        37
#> # ... with 16,486 more rows, and 1 more variable: source <chr>
#> 
#> $sources
#> # A tibble: 59 x 9
#>    source_id last_updated source source_code source_desc source_url
#>        <dbl> <date>       <chr>  <chr>       <lgl>       <lgl>     
#>  1         1 2019-10-23   Doing~ DBS         NA          NA        
#>  2         2 2019-12-20   Indic~ WDI         NA          NA        
#>  3         3 2019-11-07   Indic~ WGI         NA          NA        
#>  4         5 2016-03-21   NA     SNM         NA          NA        
#>  5         6 2019-12-02   NA     IDS         NA          NA        
#>  6        11 2013-02-22   Indic~ ADI         NA          NA        
#>  7        12 2019-12-10   Estad~ EDS         NA          NA        
#>  8        13 2017-01-03   Encue~ ESY         NA          NA        
#>  9        14 2019-12-20   Estad~ GDS         NA          NA        
#> 10        15 2019-09-13   Monit~ GEM         NA          NA        
#> # ... with 49 more rows, and 3 more variables: data_available <lgl>,
#> #   metadata_available <lgl>, concepts <dbl>
#> 
#> $topics
#> # A tibble: 21 x 3
#>    topic_id topic                topic_desc                                     
#>       <dbl> <chr>                <chr>                                          
#>  1        1 Agricultura y desar~ Para el 70% de los pobres del mundo que viven ~
#>  2        2 Eficacia de la ayuda Esta página se refiere al impacto de la asiste~
#>  3        3 Economía y crecimie~ El crecimiento de la economía constituye un as~
#>  4        4 Educación            Se trata de uno de los instrumentos más podero~
#>  5        5 Energía y minería    La economía mundial necesita cada vez más ener~
#>  6        6 Medio ambiente       Los recursos naturales y los creados por el ho~
#>  7        7 Sector financiero    Los mercados financieros son fundamentales par~
#>  8        8 Salud                Mejorar la salud es un aspecto central de los ~
#>  9        9 Infraestructura      La infraestructura determina el éxito de las a~
#> 10       10 Protección social y~ La oferta de mano de obra disponible en una ec~
#> # ... with 11 more rows
#> 
#> $regions
#> # A tibble: 48 x 4
#>    region_id iso3c iso2c region                            
#>        <dbl> <chr> <chr> <chr>                             
#>  1        NA AFR   A9     NA                               
#>  2        NA ANR   L5    "Región Andina"                   
#>  3        NA ARB   1A    "El mundo árabe"                  
#>  4        NA CAA   C9     NA                               
#>  5        NA CEA   C4     NA                               
#>  6        NA CEB   B8    "Europa Central y del Báltico\r\n"
#>  7        NA CEU   C5     NA                               
#>  8        NA CLA   C6     NA                               
#>  9        NA CME   C7     NA                               
#> 10        NA CSA   C8     NA                               
#> # ... with 38 more rows
#> 
#> $income_levels
#> # A tibble: 7 x 3
#>   iso3c iso2c income_level                  
#>   <chr> <chr> <chr>                         
#> 1 HIC   XD    Ingreso alto                  
#> 2 INX   XY    No clasificado                
#> 3 LIC   XM    Países de ingreso bajo        
#> 4 LMC   XN    Países de ingreso mediano bajo
#> 5 LMY   XO    Ingreso mediano y bajo        
#> 6 MIC   XP    Ingreso mediano               
#> 7 UMC   XT    Ingreso mediano alto          
#> 
#> $lending_types
#> # A tibble: 4 x 3
#>   iso3c iso2c lending_type                            
#>   <chr> <chr> <chr>                                   
#> 1 IBD   XF    BIRF únicamente                         
#> 2 IDB   XH    Financia-miento de  la AIF o de AIF/BIRF
#> 3 IDX   XI    AIF únicamente                          
#> 4 LNX   XX    No clasificado                          
#> 
#> $languages
#> # A tibble: 23 x 3
#>    iso2  lang       lang_native     
#>    <chr> <chr>      <chr>           
#>  1 en    English    English         
#>  2 es    Spanish    Español         
#>  3 fr    French     Français        
#>  4 ar    Arabic     <U+0639><U+0631><U+0628><U+064A>            
#>  5 zh    Chinese    <U+4E2D><U+6587>            
#>  6 bg    Bulgarian  <U+0411><U+044A><U+043B><U+0433><U+0430><U+0440><U+0441><U+043A><U+0438>       
#>  7 de    German     Deutsch         
#>  8 hi    Hindi      <U+0939><U+093F><U+0902><U+0926><U+0940>            
#>  9 id    Indonesian Bahasa indonesia
#> 10 ja    Japanese   <U+65E5><U+672C><U+8A9E>          
#> # ... with 13 more rows
#> 

# korean
wb_cache(lang = "ko")

#> $countries
#> # A tibble: 304 x 18
#>    iso3c iso2c country capital_city longitude latitude region_iso3c region_iso2c
#>    <chr> <chr> <chr>   <lgl>            <dbl>    <dbl> <chr>        <chr>       
#>  1 ABW   AW    NA      NA              -70.0     12.5  LCN          ZJ          
#>  2 AFG   AF    NA      NA               69.2     34.5  SAS          8S          
#>  3 AFR   A9    NA      NA               NA       NA    NA           NA          
#>  4 AGO   AO    NA      NA               13.2     -8.81 SSF          ZG          
#>  5 ALB   AL    NA      NA               19.8     41.3  ECS          Z7          
#>  6 AND   AD    NA      NA                1.52    42.5  ECS          Z7          
#>  7 ANR   L5    NA      NA               NA       NA    NA           NA          
#>  8 ARB   1A    NA      NA               NA       NA    NA           NA          
#>  9 ARE   AE    NA      NA               54.4     24.5  MEA          ZQ          
#> 10 ARG   AR    NA      NA              -58.4    -34.6  LCN          ZJ          
#> # ... with 294 more rows, and 10 more variables: region <chr>,
#> #   admin_region_iso3c <chr>, admin_region_iso2c <chr>, admin_region <chr>,
#> #   income_level_iso3c <chr>, income_level_iso2c <chr>, income_level <chr>,
#> #   lending_type_iso3c <chr>, lending_type_iso2c <chr>, lending_type <lgl>
#> 
#> $indicators
#> # A tibble: 16,496 x 8
#>    indicator_id indicator unit  indicator_desc source_org topics source_id
#>    <chr>        <chr>     <lgl> <chr>          <chr>      <list>     <dbl>
#>  1 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  2 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  3 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  4 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  5 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  6 1.0.HCount.~ NA        NA    NA             NA         <df[,~        37
#>  7 1.0.PGap.1.~ NA        NA    NA             NA         <df[,~        37
#>  8 1.0.PGap.2.~ NA        NA    NA             NA         <df[,~        37
#>  9 1.0.PGap.Po~ NA        NA    NA             NA         <df[,~        37
#> 10 1.0.PSev.1.~ NA        NA    NA             NA         <df[,~        37
#> # ... with 16,486 more rows, and 1 more variable: source <chr>
#> 
#> $sources
#> # A tibble: 59 x 9
#>    source_id last_updated source source_code source_desc source_url
#>        <dbl> <date>       <chr>  <chr>       <lgl>       <lgl>     
#>  1         1 2019-10-23   NA     DBS         NA          NA        
#>  2         2 2019-12-20   <U+C138><U+ACC4> <U+AC1C><U+BC1C>~ WDI         NA          NA        
#>  3         3 2019-11-07   NA     WGI         NA          NA        
#>  4         5 2016-03-21   NA     SNM         NA          NA        
#>  5         6 2019-12-02   NA     IDS         NA          NA        
#>  6        11 2013-02-22   NA     ADI         NA          NA        
#>  7        12 2019-12-10   NA     EDS         NA          NA        
#>  8        13 2017-01-03   NA     ESY         NA          NA        
#>  9        14 2019-12-20   NA     GDS         NA          NA        
#> 10        15 2019-09-13   NA     GEM         NA          NA        
#> # ... with 49 more rows, and 3 more variables: data_available <lgl>,
#> #   metadata_available <lgl>, concepts <dbl>
#> 
#> $topics
#> # A tibble: 21 x 3
#>    topic_id topic topic_desc
#>       <dbl> <lgl> <lgl>     
#>  1        1 NA    NA        
#>  2        2 NA    NA        
#>  3        3 NA    NA        
#>  4        4 NA    NA        
#>  5        5 NA    NA        
#>  6        6 NA    NA        
#>  7        7 NA    NA        
#>  8        8 NA    NA        
#>  9        9 NA    NA        
#> 10       10 NA    NA        
#> # ... with 11 more rows
#> 
#> $regions
#> # A tibble: 48 x 4
#>    region_id iso3c iso2c region
#>        <dbl> <chr> <chr> <chr> 
#>  1        NA AFR   A9    NA    
#>  2        NA ANR   L5    NA    
#>  3        NA ARB   1A    NA    
#>  4        NA CAA   C9    NA    
#>  5        NA CEA   C4    NA    
#>  6        NA CEB   B8    NA    
#>  7        NA CEU   C5    NA    
#>  8        NA CLA   C6    NA    
#>  9        NA CME   C7    NA    
#> 10        NA CSA   C8    NA    
#> # ... with 38 more rows
#> 
#> $income_levels
#> # A tibble: 7 x 3
#>   iso3c iso2c income_level
#>   <chr> <chr> <chr>       
#> 1 HIC   XD    NA          
#> 2 INX   XY    NA          
#> 3 LIC   XM    <U+C800><U+C18C><U+B4DD>      
#> 4 LMC   XN    <U+C800> <U+C911> <U+C18C><U+B4DD>  
#> 5 LMY   XO    NA          
#> 6 MIC   XP    NA          
#> 7 UMC   XT    <U+C911><U+C0C1><U+C18C><U+B4DD>    
#> 
#> $lending_types
#> # A tibble: 4 x 3
#>   iso3c iso2c lending_type
#>   <chr> <chr> <lgl>       
#> 1 IBD   XF    NA          
#> 2 IDB   XH    NA          
#> 3 IDX   XI    NA          
#> 4 LNX   XX    NA          
#> 
#> $languages
#> # A tibble: 23 x 3
#>    iso2  lang       lang_native     
#>    <chr> <chr>      <chr>           
#>  1 en    English    English         
#>  2 es    Spanish    Español         
#>  3 fr    French     Français        
#>  4 ar    Arabic     <U+0639><U+0631><U+0628><U+064A>            
#>  5 zh    Chinese    <U+4E2D><U+6587>            
#>  6 bg    Bulgarian  <U+0411><U+044A><U+043B><U+0433><U+0430><U+0440><U+0441><U+043A><U+0438>       
#>  7 de    German     Deutsch         
#>  8 hi    Hindi      <U+0939><U+093F><U+0902><U+0926><U+0940>            
#>  9 id    Indonesian Bahasa indonesia
#> 10 ja    Japanese   <U+65E5><U+672C><U+8A9E>          
#> # ... with 13 more rows
#> 

# possible values of lang are in the iso2 column in the languages data frame
example_iso2 <- wb_languages()$iso2[1]
wb_cache(lang = example_iso2)

#> $countries
#> # A tibble: 304 x 18
#>    iso3c iso2c country capital_city longitude latitude region_iso3c region_iso2c
#>    <chr> <chr> <chr>   <chr>            <dbl>    <dbl> <chr>        <chr>       
#>  1 ABW   AW    Aruba   Oranjestad      -70.0     12.5  LCN          ZJ          
#>  2 AFG   AF    Afghan~ Kabul            69.2     34.5  SAS          8S          
#>  3 AFR   A9    Africa  NA               NA       NA    NA           NA          
#>  4 AGO   AO    Angola  Luanda           13.2     -8.81 SSF          ZG          
#>  5 ALB   AL    Albania Tirane           19.8     41.3  ECS          Z7          
#>  6 AND   AD    Andorra Andorra la ~      1.52    42.5  ECS          Z7          
#>  7 ANR   L5    Andean~ NA               NA       NA    NA           NA          
#>  8 ARB   1A    Arab W~ NA               NA       NA    NA           NA          
#>  9 ARE   AE    United~ Abu Dhabi        54.4     24.5  MEA          ZQ          
#> 10 ARG   AR    Argent~ Buenos Aires    -58.4    -34.6  LCN          ZJ          
#> # ... with 294 more rows, and 10 more variables: region <chr>,
#> #   admin_region_iso3c <chr>, admin_region_iso2c <chr>, admin_region <chr>,
#> #   income_level_iso3c <chr>, income_level_iso2c <chr>, income_level <chr>,
#> #   lending_type_iso3c <chr>, lending_type_iso2c <chr>, lending_type <chr>
#> 
#> $indicators
#> # A tibble: 16,496 x 8
#>    indicator_id indicator unit  indicator_desc source_org topics source_id
#>    <chr>        <chr>     <lgl> <chr>          <chr>      <list>     <dbl>
#>  1 1.0.HCount.~ Poverty ~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  2 1.0.HCount.~ Poverty ~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  3 1.0.HCount.~ Middle C~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  4 1.0.HCount.~ Official~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  5 1.0.HCount.~ Poverty ~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  6 1.0.HCount.~ Vulnerab~ NA    The poverty h~ LAC Equit~ <df[,~        37
#>  7 1.0.PGap.1.~ Poverty ~ NA    The poverty g~ LAC Equit~ <df[,~        37
#>  8 1.0.PGap.2.~ Poverty ~ NA    The poverty g~ LAC Equit~ <df[,~        37
#>  9 1.0.PGap.Po~ Poverty ~ NA    The poverty g~ LAC Equit~ <df[,~        37
#> 10 1.0.PSev.1.~ Poverty ~ NA    The poverty s~ LAC Equit~ <df[,~        37
#> # ... with 16,486 more rows, and 1 more variable: source <chr>
#> 
#> $sources
#> # A tibble: 59 x 9
#>    source_id last_updated source source_code source_desc source_url
#>        <dbl> <date>       <chr>  <chr>       <lgl>       <lgl>     
#>  1         1 2019-10-23   Doing~ DBS         NA          NA        
#>  2         2 2019-12-20   World~ WDI         NA          NA        
#>  3         3 2019-11-07   World~ WGI         NA          NA        
#>  4         5 2016-03-21   Subna~ SNM         NA          NA        
#>  5         6 2019-12-02   Inter~ IDS         NA          NA        
#>  6        11 2013-02-22   Afric~ ADI         NA          NA        
#>  7        12 2019-12-10   Educa~ EDS         NA          NA        
#>  8        13 2017-01-03   Enter~ ESY         NA          NA        
#>  9        14 2019-12-20   Gende~ GDS         NA          NA        
#> 10        15 2019-09-13   Globa~ GEM         NA          NA        
#> # ... with 49 more rows, and 3 more variables: data_available <lgl>,
#> #   metadata_available <lgl>, concepts <dbl>
#> 
#> $topics
#> # A tibble: 21 x 3
#>    topic_id topic                topic_desc                                     
#>       <dbl> <chr>                <chr>                                          
#>  1        1 Agriculture & Rural~ "For the 70 percent of the world's poor who li~
#>  2        2 Aid Effectiveness    "Aid effectiveness is the impact that aid has ~
#>  3        3 Economy & Growth     "Economic growth is central to economic develo~
#>  4        4 Education            "Education is one of the most powerful instrum~
#>  5        5 Energy & Mining      "The world economy needs ever-increasing amoun~
#>  6        6 Environment          "Natural and man-made environmental resources ~
#>  7        7 Financial Sector     "An economy's financial markets are critical t~
#>  8        8 Health               "Improving health is central to the Millennium~
#>  9        9 Infrastructure       "Infrastructure helps determine the success of~
#> 10       10 Social Protection &~ "The supply of labor available in an economy i~
#> # ... with 11 more rows
#> 
#> $regions
#> # A tibble: 48 x 4
#>    region_id iso3c iso2c region                                              
#>        <dbl> <chr> <chr> <chr>                                               
#>  1        NA AFR   A9    Africa                                              
#>  2        NA ANR   L5    Andean Region                                       
#>  3        NA ARB   1A    Arab World                                          
#>  4        NA CAA   C9    Sub-Saharan Africa (IFC classification)             
#>  5        NA CEA   C4    East Asia and the Pacific (IFC classification)      
#>  6        NA CEB   B8    Central Europe and the Baltics                      
#>  7        NA CEU   C5    Europe and Central Asia (IFC classification)        
#>  8        NA CLA   C6    Latin America and the Caribbean (IFC classification)
#>  9        NA CME   C7    Middle East and North Africa (IFC classification)   
#> 10        NA CSA   C8    South Asia (IFC classification)                     
#> # ... with 38 more rows
#> 
#> $income_levels
#> # A tibble: 7 x 3
#>   iso3c iso2c income_level       
#>   <chr> <chr> <chr>              
#> 1 HIC   XD    High income        
#> 2 INX   XY    Not classified     
#> 3 LIC   XM    Low income         
#> 4 LMC   XN    Lower middle income
#> 5 LMY   XO    Low & middle income
#> 6 MIC   XP    Middle income      
#> 7 UMC   XT    Upper middle income
#> 
#> $lending_types
#> # A tibble: 4 x 3
#>   iso3c iso2c lending_type  
#>   <chr> <chr> <chr>         
#> 1 IBD   XF    IBRD          
#> 2 IDB   XH    Blend         
#> 3 IDX   XI    IDA           
#> 4 LNX   XX    Not classified
#> 
#> $languages
#> # A tibble: 23 x 3
#>    iso2  lang       lang_native     
#>    <chr> <chr>      <chr>           
#>  1 en    English    English         
#>  2 es    Spanish    Español         
#>  3 fr    French     Français        
#>  4 ar    Arabic     <U+0639><U+0631><U+0628><U+064A>            
#>  5 zh    Chinese    <U+4E2D><U+6587>            
#>  6 bg    Bulgarian  <U+0411><U+044A><U+043B><U+0433><U+0430><U+0440><U+0441><U+043A><U+0438>       
#>  7 de    German     Deutsch         
#>  8 hi    Hindi      <U+0939><U+093F><U+0902><U+0926><U+0940>            
#>  9 id    Indonesian Bahasa indonesia
#> 10 ja    Japanese   <U+65E5><U+672C><U+8A9E>          
#> # ... with 13 more rows
#>