Search indicator information available through the World Bank API

This function allows finds indicators that match a search term and returns a data frame of matching results

wb_search(
  pattern,
  fields = c("indicator_id", "indicator", "indicator_desc"),
  extra = FALSE,
  cache,
  ignore.case = TRUE,
  ...
)

Arguments

pattern
Character string or regular expression to be matched

fields
Character vector of column names through which to search

extra
if FALSE, only the indicator ID and short name are returned, if TRUE, all columns of the cache parameter's indicators data frame are returned. Default is FALSE

cache
List of data frames returned from wb_cache(). If omitted, wb_cachelist is used

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching. Default is TRUE

...
Any additional grep() agruments you which to pass

Value

A tibble with indicators that match the search pattern.

Examples

wb_search(pattern = "education")

#> # A tibble: 2,889 x 3
#>    indicator_id     indicator                indicator_desc                     
#>    <chr>            <chr>                    <chr>                              
#>  1 2.0.hoi.Cel      HOI: Mobile Phone        The Human Opportunities Index (HOI~
#>  2 2.0.hoi.Ele      HOI: Electricity         The Human Opportunities Index (HOI~
#>  3 2.0.hoi.FPS      HOI: Finished Primary S~ The Human Opportunities Index (HOI~
#>  4 2.0.hoi.Int      HOI: Internet            The Human Opportunities Index (HOI~
#>  5 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#>  6 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#>  7 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#>  8 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#>  9 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#> 10 2.0.hoi.Math.pl~ HOI: Mathematics Profic~ The Human Opportunities Index (HOI~
#> # ... with 2,879 more rows

wb_search(pattern = "Food and Agriculture Organization", fields = "source_org")

#> # A tibble: 131 x 3
#>    indicator_id   indicator                  indicator_desc                     
#>    <chr>          <chr>                      <chr>                              
#>  1 AG.AGR.TRAC.NO Agricultural machinery, t~ Agricultural machinery refers to t~
#>  2 AG.CON.FERT.P~ Fertilizer consumption (%~ Fertilizer consumption measures th~
#>  3 AG.CON.FERT.ZS Fertilizer consumption (k~ Fertilizer consumption measures th~
#>  4 AG.CRP.BLY.CD  Producer Price for Barley~ Producer prices are prices receive~
#>  5 AG.CRP.BLY.CN  Producer Price for Barley~ Producer prices are prices receive~
#>  6 AG.CRP.FNO.CD  Producer Price for Fonio ~ Producer prices are prices receive~
#>  7 AG.CRP.FNO.CN  Producer Price for Fonio ~ Producer prices are prices receive~
#>  8 AG.CRP.MLT.CD  Producer Price for Millet~ Producer prices are prices receive~
#>  9 AG.CRP.MLT.CN  Producer Price for Millet~ Producer prices are prices receive~
#> 10 AG.CRP.MZE.CD  Producer Price for Maize ~ Producer prices are prices receive~
#> # ... with 121 more rows

# with regular expression operators
# 'poverty' OR 'unemployment' OR 'employment'
wb_search(pattern = "poverty|unemployment|employment")

#> # A tibble: 829 x 3
#>    indicator_id     indicator             indicator_desc                        
#>    <chr>            <chr>                 <chr>                                 
#>  1 1.0.HCount.1.90~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  2 1.0.HCount.2.5u~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  3 1.0.HCount.Mid1~ Middle Class ($10-50~ The poverty headcount index measures ~
#>  4 1.0.HCount.Ofcl  Official Moderate Po~ The poverty headcount index measures ~
#>  5 1.0.HCount.Poor~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  6 1.0.HCount.Vul4~ Vulnerable ($4-10 a ~ The poverty headcount index measures ~
#>  7 1.0.PGap.1.90usd Poverty Gap ($1.90 a~ The poverty gap captures the mean agg~
#>  8 1.0.PGap.2.5usd  Poverty Gap ($2.50 a~ The poverty gap captures the mean agg~
#>  9 1.0.PGap.Poor4u~ Poverty Gap ($4 a da~ The poverty gap captures the mean agg~
#> 10 1.0.PSev.1.90usd Poverty Severity ($1~ The poverty severity index combines i~
#> # ... with 819 more rows

# pass any other grep argument along as well
# everything without 'education'
wb_search(pattern = "education", invert = TRUE)

#> # A tibble: 16,496 x 3
#>    indicator_id     indicator             indicator_desc                        
#>    <chr>            <chr>                 <chr>                                 
#>  1 1.0.HCount.1.90~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  2 1.0.HCount.2.5u~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  3 1.0.HCount.Mid1~ Middle Class ($10-50~ The poverty headcount index measures ~
#>  4 1.0.HCount.Ofcl  Official Moderate Po~ The poverty headcount index measures ~
#>  5 1.0.HCount.Poor~ Poverty Headcount ($~ The poverty headcount index measures ~
#>  6 1.0.HCount.Vul4~ Vulnerable ($4-10 a ~ The poverty headcount index measures ~
#>  7 1.0.PGap.1.90usd Poverty Gap ($1.90 a~ The poverty gap captures the mean agg~
#>  8 1.0.PGap.2.5usd  Poverty Gap ($2.50 a~ The poverty gap captures the mean agg~
#>  9 1.0.PGap.Poor4u~ Poverty Gap ($4 a da~ The poverty gap captures the mean agg~
#> 10 1.0.PSev.1.90usd Poverty Severity ($1~ The poverty severity index combines i~
#> # ... with 16,486 more rows

# contains "gdp" AND "trade"
wb_search("^(?=.*gdp)(?=.*trade).*", perl = TRUE)

#> # A tibble: 12 x 3
#>    indicator_id   indicator                     indicator_desc                  
#>    <chr>          <chr>                         <chr>                           
#>  1 BG.GSR.NFSV.G~ Trade in services (% of GDP)  Trade in services is the sum of~
#>  2 CM.MKT.TRAD.G~ Stocks traded, total value (~ The value of shares traded is t~
#>  3 EF.EFM.OVRL.XD Economic Fitness Metric       Economic Fitness (EF) is both a~
#>  4 EF.EFM.PROD.XD Economic Fitness Metric       Economic Fitness (EF) is both a~
#>  5 EF.EFM.UNIV.XD Universal Economic Fitness M~ The Universal Economic Fitness ~
#>  6 GFDD.DM.02     Stock market total value tra~ Total value of all traded share~
#>  7 NA.GDP.TRD.HT~ GDP on Trade, Hotel and Rest~ NA                              
#>  8 NA.GDP.TRD.HT~ GDP on Trade, Hotel and Rest~ NA                              
#>  9 NE.MRCH.GDP.ZS Merchandise trade to GDP rat~ Merchandise trade as a share of~
#> 10 NE.TRD.GNFS.ZS Trade (% of GDP)              Trade is the sum of exports and~
#> 11 NY.GDY.TOTL.KN Gross domestic income (const~ Gross domestic income is derive~
#> 12 TG.VAL.TOTL.G~ Merchandise trade (% of GDP)  Merchandise trade as a share of~

# contains "gdp" and NOT "trade"
wb_search("^(?=.*gdp)(?!.*trade).*", perl = TRUE)

#> # A tibble: 485 x 3
#>    indicator_id    indicator                    indicator_desc                  
#>    <chr>           <chr>                        <chr>                           
#>  1 5.51.01.10.gdp  Per capita GDP growth        GDP per capita is the sum of gr~
#>  2 6.0.GDP_current GDP (current $)              GDP is the sum of gross value a~
#>  3 6.0.GDP_growth  GDP growth (annual %)        Annual percentage growth rate o~
#>  4 6.0.GDP_usd     GDP (constant 2005 $)        GDP is the sum of gross value a~
#>  5 6.0.GDPpc_cons~ GDP per capita, PPP (consta~ GDP per capita based on purchas~
#>  6 BI.WAG.TOTL.GD~ Wage bill as a percentage o~ NA                              
#>  7 BM.KLT.DINV.WD~ Foreign direct investment, ~ Foreign direct investment refer~
#>  8 BN.CAB.XOKA.GD~ Current account balance (% ~ Current account balance is the ~
#>  9 BN.CUR.GDPM.ZS  Current account balance exc~ Current account balance is the ~
#> 10 BN.GSR.FCTY.CD~ Net income (% of GDP)        Net income refers to receipts a~
#> # ... with 475 more rows